+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog for Venstar ColorTouch Thermostat Hubitat Driver
 
+## 21270
+
+* Add config to optinally force device units (default: use hub's configured unit)
+* Improve error handling for certain exceptions during request
+* Temperature unit default for T2000
+* May support T5800/5900, but I can't test since I don't have one.
+
 ## 21268
 
 * Clean up HTTP Digest authentication; add state variable indicator for which is used (debug assist);